rename global to store
